added config for matter overdrive compat and moved matter overdrive check class
